INTERLOK-413 #comment Updated tests. INTERLOK-413 #time 1h
Due to the change log for java8 collections (http://docs.oracle.com/javase/8/docs/technotes/guides/collections/changes8.html)
"In rare situations, this change could introduce a change
to the iteration order of HashMap and HashSet. A particular iteration
order is not specified for HashMap objects - any code that depends on
iteration order should be fixed."
That statement has burned us, so we have to switch how we actually test
the results rather than doing a raw comparison.
The ordering of the XML and JSON elements may change.