INTERLOK-758 If the string doesn't parse as a JSONObject, try JSONArray INTERLOK-758 #time 2h
XML.toString() doesn't create "valid" XML with a JSONArray object.
It effectively creates the string
<json.../>
<json.../>
<json.../>
Which won't work if you try to do a transform; So, if the object
parses as a JSONArray, we need to wrap it as a JSONObject so effectively
we get (this isn't ideal, but means we can use the SimpleTransformDriver
to get it into XML).
<json>
  <array-item.../>
  <array-item.../>
  <array-item.../>
</json>