INTERLOK-1499 Fix XMLSerializer so that empty elements are not an array. INTERLOK-1499 #time 1h
If getChildCount = 0 && getChildElements.size() == 0; then its not treated as
an array. This has the side effect that <xxx type="String"/> will become
{"xxx" : null}} rather than "xxx": "".
Two workarounds for this : don't use DefaultTransformationDriver; the
simple transformation driver will omit "xxx" entirely if it was originally used
to create the json-xml.
The other is to use a jolt-transform "c.a.c.t.j.j.NullToEmptyString".